LayoutTests/js/dfg-to-string-toString-becomes-bad-with-check-structure-expected.txt

Tests that the DFG checks that the toString method didn't become bad even if the StringObject already had a CheckStructure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "foo"
PASS foo.call(new String("foo")) is "42"
PASS successfullyParsed is true

TEST COMPLETE
